20th commit, another update

+++ public/js/blog.js
@@ -1,19 +1,22 @@
+const deleteButton = document.getElementById("deleteBlog");
+const blogId = document.getElementById("datePosted").dataset.blogid;
 const handleDeleteBlog = async (event) => {
   event.preventDefault();
 
   const id = document
     .querySelector("[data-blogID]")
     .getAttribute("data-blogID");
+  if (id) {
+    const delResponse = await fetch(`/api/blog/${blogId}`, {
+      method: "DELETE",
+    });
 
-  const response = await fetch(`/api/blog/${id}`, {
-    method: "DELETE",
-  });
-
-  if (response.ok) {
-    //if successful, redirect browser to home page
-    document.location.replace("/");
-    alert("Blog successfully deleted!");
+    if (delResponse.ok) {
+      //if successful, redirect browser to home page
+      document.location.replace("/");
+      alert("Blog successfully deleted!");
+    }
   }
 };
 
-document.addEventListener("click", handleDeleteBlog);
+deleteButton.addEventListener("click", handleDeleteBlog);
